Able to run download and mutation classifier notebook
Updated:
- dependencies & Dockerfile: since the worker is running python
notebooks from the cognoma/machine-learning repo, it needs the conda
dependencies from that repo. Creating a conda environment in the docker
container created a large container of 2.6 gigabytes. Installing those
dependencies directly with pip shaved the size to 1.05 gigabytes. Due
to using pip we will need to keep an eye out to make sure dependencies
and environments are all in sync and working. If issues arise we can
revert to creating the conda environment rather than using pip.
- Dockerfile continued: running the docker container now will run
ml_task_runner.py. In it’s current state, that means running the
download notebook (notebook 1) and the mutation classifier notebook (#2)
- ml_task_runner.py: the worker is still unable to integrate with
task-service. Consequently I commented out that code so that the script
can just run the notebooks as is for testing.
- api_clients: fixed errors with using the requests library. Added
handling for non-200 responses.