Refactored disease slicing and fixed slicing error
The data frames were being sliced but not stored back into their
original variables, they were just being printed. This was causing
obvious errors because the dimensions of the data frames wouldn’t line
up.

On advice of @dhimmel I refactored the way samples are sliced by
disease by taking a row-wise max of the relevant disease acronym
columns and filtering on whether that max is greater than 1 (the sample
has one of the specified diseases).